[RUMF-828] split internalMonitoring
To avoid internalMonitoring cyclic dependencies, split it in three
files:

* `configuration` holds the monitoring configuration and batch
reference. It is dependency free and shared between the other files
* `start` holds all functions to configure internal monitoring and
start the batch process
* `monitor` holds all functions to actually monitor executed code
