+++ packages/core/src/domain/internalMonitoring/configuration.ts
@@ -1,0 +1,8 @@
+import type { Batch } from '../../transport/transport'
+
+export const monitoringConfiguration: {
+  batch?: Batch
+  debugMode?: boolean
+  maxMessagesPerPage: number
+  sentMessageCount: number
+} = { maxMessagesPerPage: 0, sentMessageCount: 0 }
