+++ packages/core/src/domain/internalMonitoring/monitor.ts
@@ -1,0 +1,121 @@
+import type { Context } from '../../tools/context'
+import { toStackTraceString } from '../../tools/error'
+import { ThisParameterType, Parameters, jsonStringify } from '../../tools/utils'
+import { computeStackTrace } from '../tracekit/computeStackTrace'
+import { monitoringConfiguration } from './configuration'
+
+export enum StatusType {
+  info = 'info',
+  error = 'error',
+}
+
+export interface MonitoringMessage extends Context {
+  message: string
+  status: StatusType
+  error?: {
+    kind?: string
+    stack: string
+  }
+}
+
+export function monitored<T extends (...params: any[]) => unknown>(
+  _: any,
+  __: string,
+  descriptor: TypedPropertyDescriptor<T>
+) {
+  const originalMethod = descriptor.value!
+  descriptor.value = function (this: any, ...args: Parameters<T>) {
+    const decorated = monitoringConfiguration.batch ? monitor(originalMethod) : originalMethod
+    return decorated.apply(this, args) as ReturnType<T>
+  } as T
+}
+
+export function monitor<T extends (...args: any[]) => any>(fn: T): T {
+  return (function (this: any) {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+    return callMonitored(fn, this, (arguments as unknown) as Parameters<T>)
+  } as unknown) as T // consider output type has input type
+}
+
+export function callMonitored<T extends (...args: any[]) => any>(
+  fn: T,
+  context: ThisParameterType<T>,
+  args: Parameters<T>
+): ReturnType<T> | undefined
+export function callMonitored<T extends (this: void) => any>(fn: T): ReturnType<T> | undefined
+export function callMonitored<T extends (...args: any[]) => any>(
+  fn: T,
+  context?: any,
+  args?: any
+): ReturnType<T> | undefined {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+    return fn.apply(context, args)
+  } catch (e) {
+    logErrorIfDebug(e)
+    try {
+      addErrorToMonitoringBatch(e)
+    } catch (e) {
+      logErrorIfDebug(e)
+    }
+  }
+}
+
+export function addMonitoringMessage(message: string, context?: Context) {
+  logMessageIfDebug(message)
+  addToMonitoringBatch({
+    message,
+    ...context,
+    status: StatusType.info,
+  })
+}
+
+export function addErrorToMonitoringBatch(e: unknown) {
+  addToMonitoringBatch({
+    ...formatError(e),
+    status: StatusType.error,
+  })
+}
+
+function addToMonitoringBatch(message: MonitoringMessage) {
+  if (
+    monitoringConfiguration.batch &&
+    monitoringConfiguration.sentMessageCount < monitoringConfiguration.maxMessagesPerPage
+  ) {
+    monitoringConfiguration.sentMessageCount += 1
+
+    monitoringConfiguration.batch.add(message)
+  }
+}
+
+function formatError(e: unknown) {
+  if (e instanceof Error) {
+    const stackTrace = computeStackTrace(e)
+    return {
+      error: {
+        kind: stackTrace.name,
+        stack: toStackTraceString(stackTrace),
+      },
+      message: stackTrace.message!,
+    }
+  }
+  return {
+    error: {
+      stack: 'Not an instance of error',
+    },
+    message: `Uncaught ${jsonStringify(e)!}`,
+  }
+}
+
+function logErrorIfDebug(e: any) {
+  if (monitoringConfiguration.debugMode) {
+    // Log as warn to not forward the logs.
+    console.warn('[INTERNAL ERROR]', e)
+  }
+}
+
+function logMessageIfDebug(message: any) {
+  if (monitoringConfiguration.debugMode) {
+    console.log('[MONITORING MESSAGE]', message)
+  }
+}
